Exceptions

+++ src/main/java/com/desafio_3/services/ClientService.java
@@ -1,15 +1,21 @@
 package com.desafio_3.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.desafio_3.dto.ClientDTO;
 import com.desafio_3.entities.Client;
 import com.desafio_3.mappers.ClientMapper;
 import com.desafio_3.repositories.ClientRepository;
+import com.desafio_3.services.exceptions.DatabaseException;
+import com.desafio_3.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
@@ -19,7 +25,7 @@ public class ClientService {
 	
 	@Transactional(readOnly = true)
 	public ClientDTO findById(Long id) {
-		Client client = repository.findById(id).get();
+		Client client = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return ClientMapper.INSTANCE.toDTO(client);
 	}
 	
@@ -39,15 +45,27 @@ public class ClientService {
 	
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO dto) {
-		Client client = repository.getReferenceById(id);
-		ClientMapper.INSTANCE.updateFromDTO(dto, client);
-		Client updatedClient = repository.save(client);
-		return ClientMapper.INSTANCE.toDTO(updatedClient);
+		try {
+			Client client = repository.getReferenceById(id);
+			ClientMapper.INSTANCE.updateFromDTO(dto, client);
+			Client updatedClient = repository.save(client);
+			return ClientMapper.INSTANCE.toDTO(updatedClient);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso Não encontrado");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(long id) {
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
+		
 	}
 	
 }
